requirements.txt
kafka-python==2.0.2
cassandra-driver==3.29.1
pandas==2.2.2
kaggle==1.6.17
pyspark==3.5.1
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
pyarrow==16.1.0
